Ping type changed

+++ JChat/src/com/intgames/JChat/core/Server.java
@@ -63,7 +63,7 @@ public class Server {
 		
 	}
 	
-	public void sendEveryone(Message msg, double ping) {
+	public void sendEveryone(Message msg, long ping) {
 		
 		log.println(msg, ping);
